% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_data.R
\name{addtimecols}
\alias{addtimecols}
\title{Extract information on timing of events from pasted-together columns}
\usage{
addtimecols(data, times, split = "\\\\|", from_suffix = "_from_index")
}
\arguments{
\item{data}{data from mk_data}

\item{times}{a list of lists. See details}

\item{split}{character that separates times in pasted cells (default = "\\|")}

\item{from_suffix}{suffix of columns containing relative times pasted. default = "from_index"}
}
\value{
modified data w/ additional columns
}
\description{
mk_data pastes the number of units time between the 'index' event
and a given event (ICD/CPT/etc) to a single cell separated by "|" .
e.g. "-45|0|90" would be an event that occured 45 before, concurrent,
and 90 days after index. This allows the contruction of columns from these cells.
}
\details{
times should be a list of lists such that each element is an event of interest
and the sublist contains the times of interest. Using Inf or -Inf in the range of
times will return the closests event after or before the other end of the range (which must be non-Inf).
For example times = list(eventA = list(c(0,90), c(-Inf, -1)))
would result in a column containing T/F for if eventA occurs in the interval \link{0, 90}
and a column containing the closest event to the day before index(inclusive). Note
eventA_from_index (the column containing the pasted times) must exist in data.
}
