% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_codes.R
\name{count_codes}
\alias{count_codes}
\title{Count the number of times particular codes are matched}
\usage{
count_codes(codes, code_list = NULL, sep = " \\\\| ")
}
\arguments{
\item{codes}{vector of codes from n encounters}

\item{code_list}{list of ICD/CPT/charge codes of interest}

\item{sep}{character separating individual codes in \code{codes} (default = " \\| ")}
}
\value{
data.frame containing counts of individual codes (\code{group} = name in code_list)
}
\description{
Count the number of times particular codes are matched
}
