% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_data.R
\name{get_comorbidities}
\alias{get_comorbidities}
\title{Wrapper for comorbidity package allowing for mix of 9 and 10}
\usage{
get_comorbidities(data, type = c("charlson", "elixhauser"))
}
\arguments{
\item{data}{with columns MEDREC_KEY, ICD_CODE, ICD_VERSION - where
the latter 2 have values pasted together separated by "|"}

\item{type}{"charlson" or "elixhauser"}
}
\value{
data.frame of individual comorbidities (T/F) and the index
}
\description{
Wrapper for comorbidity package allowing for mix of 9 and 10
}
\keyword{internal}
