% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_encounters_within_mrange.R
\name{get_encounters_within_mrange}
\alias{get_encounters_within_mrange}
\title{Find encounters within a particular time-frame}
\usage{
get_encounters_within_mrange(
  data,
  cols,
  month_col,
  num_months = c(0, 3),
  by = c("MEDREC_KEY"),
  fun = function(x) any(x == 1),
  add_columns_to_data = T,
  exclude_col = NULL
)
}
\arguments{
\item{data}{from \code{read_premier}}

\item{cols}{columns of interest}

\item{month_col}{column containing number of months from event of interest}

\item{num_months}{a vector of length 2 with lower and upper bounds (inclusive)}

\item{by}{what to aggregate over (default = "MEDREC_KEY")}

\item{fun}{function to apply (default = function(x) any(x == 1))}

\item{add_columns_to_data}{add the new columns to \code{data}?}

\item{exclude_col}{column name to exclude when looking (used for excluding the index encounter)}
}
\value{
Either the original data with new columns (from \code{cols}) identifying those with a match within
the time frame specified by \code{num_months} (otherwise just those columns if \code{add_columns_to_data = F})
}
\description{
Find encounters within a particular time-frame
}
