% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cols.R
\name{make_cols}
\alias{make_cols}
\title{Create columns signifying a code match}
\usage{
make_cols(codes, match_list = NULL, codes_matched = NULL)
}
\arguments{
\item{codes}{list of codes of interest (passed through \code{sanitize_codes})}

\item{match_list}{codes associated with hospital encounters in list of vectors form}

\item{codes_matched}{vector of codes associated with hospital encounters separated by " | ".
This is needed if \code{match_list} is not provided, in which case \code{match_list = strsplit(codes_matched, split = " \\| ")}}
}
\value{
a data.frame with \code{length(codes)} columns and \code{length(match_list)} rows with a
0 if no codes in \code{codes} were identified for an encounter and a 1 if there was at least one match.
}
\description{
Create columns signifying a code match
}
