% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_data.R
\name{mk_data}
\alias{mk_data}
\title{Combine PHD data}
\usage{
mk_data(
  files,
  index_event,
  medrec_keys,
  combine,
  comorbidity = c("none", "charlson", "elixhauser"),
  comorbidity_time = c(-365, -1),
  ignore_from_index,
  return_all = FALSE
)
}
\arguments{
\item{files}{vector of files to read in and combine}

\item{index_event}{character. Column containing event of interest
(the first instance of which will be called index). Can be created in combine}

\item{medrec_keys}{optional vector if IDs to subset on}

\item{combine}{named list of variables to create that are combinations
of existing variables. Best to use alist. e.g. combine = alist("A" = cpt_A | icdp_A)}

\item{comorbidity}{"none", "charlson", or "elixhauser". Uses the comorbidity package
on ICD codes within a number of days of index defined by comorbidity_time}

\item{comorbidity_time}{time window to use ICD codes to calculate comorbidity index from}

\item{ignore_from_index}{optional vector of columns to ignore when pasting time from index
(e.g. if interest is only presence at index)}

\item{return_all}{return all events or just index (default)}
}
\value{
data.table of index events (if return_all = F) with
comorbidity info (if requested) and columns containing time from index
for events defined by ICD, CPT, or premier charge codes.
}
\description{
For running on the results of calls to read_premier2
to calculate days between index_event and events
defined by ICD, CPT, charge codes (identified via column prefix)
}
