% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_premier2.R
\name{read_premier2}
\alias{read_premier2}
\title{Read newer premier data}
\usage{
read_premier2(
  year,
  quarter,
  cpt,
  icd9_diag,
  icd10_diag,
  icd9_proc,
  icd10_proc,
  charge,
  medrec_keys,
  pat_keys,
  file_prefix = "uhc_",
  dir = getwd(),
  partial = FALSE,
  cols_to_keep = list(cpt = c(), icd_diag = c(), icd_proc = c(), charge = c())
)
}
\arguments{
\item{year}{of data (we have 2000 - 2022)}

\item{quarter}{1-4}

\item{cpt}{named list of CPT codes}

\item{icd9_diag}{named list}

\item{icd10_diag}{named list}

\item{icd9_proc}{named list}

\item{icd10_proc}{named list}

\item{charge}{named list of premier charge codes}

\item{medrec_keys}{vector of MEDREC_KEY (patients) to filter on}

\item{pat_keys}{vector of PAT_KEY (encounters) to filter on}

\item{file_prefix}{start of file names}

\item{dir}{where the files are}

\item{partial}{are we looking for partial code matches for matching sub-codes (e.g. 188.1 matched by 188.12) or literal matches (probably faster)}

\item{cols_to_keep}{list of additional columns from the various PHD tables.
e.g. if you edited charges = "BILL_CHARGES", the result would have a column for BILL_CHARGES with each element associated with an encounter pasted
together separated by "|". By default this only happens for the "_CODE" columns (and "_VERSION" for ICD)}
}
\value{
data.table of results from PATDEMO merged with info from other tables
depending on the code systems specified.
}
\description{
Read newer premier data
}
