% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_premier_codes.R
\name{read_premier_codes}
\alias{read_premier_codes}
\title{Read specific ICD, CPT, and charge codes from premier data files}
\usage{
read_premier_codes(
  files,
  icd9 = NULL,
  icd10 = NULL,
  cpt = NULL,
  charge = NULL,
  pat_keys = NULL,
  pat_filter_only = F
)
}
\arguments{
\item{files}{Character vector of PHD files containing ICD diagnosis, ICD procedure, CPT/HCPS, or premier charge/billing codes (all of same type).
Assumed to be SAS files (read by \code{haven::read_sas})}

\item{icd9}{named list of ICD 9 codes of interest (whether procedure or diagnosis is determined by \code{files})}

\item{icd10}{named list of ICD 10 codes of interest}

\item{cpt}{named list of CPT codes of interest}

\item{charge}{named list of premier charge codes of interest (only one of icd9, icd10, cpt, charge should be provided)}

\item{pat_keys}{vector of PAT_KEY encounter IDs to keep. If icd9=icd10=cpt=charge=NULL and PAT_KEY is specified, the function will resturn all}

\item{pat_filter_only}{should encounters only be kept if they appear in pat_keys?}
}
\value{
Either the original data with new columns (from \code{cols}) identifying those with a match within
the time frame specified by \code{num_months} (otherwise just those columns if \code{add_columns_to_data = F})
}
\description{
Read specific ICD, CPT, and charge codes from premier data files
}
