% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_codes.R
\name{sanitize_codes}
\alias{sanitize_codes}
\title{Modify codes to allow for inclusive matching}
\usage{
sanitize_codes(code_list, char = "x")
}
\arguments{
\item{code_list}{list of ICD or CPT codes}

\item{x}{the character used to allow for inclusive matching (default = "x").
For example, to match codes "99.\link{0:10}" (and all sub-codes, like 99.01), one would
put "99.x" in their code list. This would be translated to "99." for inclusive matching.
Codes not ending in \code{char} have $ put at the end to stop matching.}
}
\value{
modified version of code_list to allow for inclusive matching where necessary
}
\description{
Modify codes to allow for inclusive matching
}
